Non-stationary Maximal-overlapping Allan Variance

Calculation of the theoretical Maximal-overlapping Allan variance for constant-mean non-stationary time series data.

MOAV(n, covmat)

Arguments

n
An integer indicating the length of each vector of consecutive observations considered for the average.

covmat
A matrix indicating the T-by-T covariance matrix of the time series with length T.

Value

A field <numeric> that is the theoretical Maximal-overlapping Allan variance for constant-mean non-stationary time series data.

Details

This calculation of Maximal-overlapping Allan variance is based on the definition on "A Study of the Allan Variance for Constant-Mean Non-Stationary Processes" by Xu et al., 2017, IEEE Signal Processing Letters, 24(8): 12571260. Here n is an integer larger than 1 and smaller than \(floor\left(log_2 \left(dim\left(covmat\right)[1]\right)\right)-1\).

Examples

library(simts)
set.seed(999)
x = gen_gts(AR1(phi = 0.3, sigma2 = 1), n = 1000, freq = 1)
avar(x, type = "to")

#> 
#>  Clusters: 
#> [1]   2   4   8  16  32  64 128 256
#> 
#>  Allan Variances: 
#> [1] 0.4564547 0.4862027 0.2664077 0.1376303 0.0423879 0.0361422 0.0081693
#> [8] 0.0070663
#> 
#>  Errors: 
#> [1] 0.031654 0.044811 0.063500 0.090167 0.128565 0.184900 0.270914 0.414781

a = matrix(rep(0, 1000^2), nrow = 1000)
for (i in 1:1000){
  a[,i] = seq(from = 1 - i, length.out = 1000)
}
a.diag = diag(a)
a[upper.tri(a,diag=TRUE)] = 0
a = a + t(a) + diag(a.diag)
covmat = 0.3^a
sapply(1:8, function(y){MOAV(2^y, covmat)})

#> [1] 0.523250000 0.350727081 0.203446388 0.108896684 0.056242028 0.028569436
#> [7] 0.014396823 0.007226438