Internal function to the Allan Variance Linear Regression estimator

Estimate the parameters of time series models based on the Allan Variance Linear Regression (AVLR) approach

fit_avlr(qn, wn, rw, dr, ad, scales)

Arguments

qn
A vec specifying on which scales the parameters of a Quantization Noise (QN) should be computed.

wn
A vec specifying on which scales the parameters of a White Noise (WN) should be computed.

rw
A vec specifying on which scales the parameters of a Random Wakk (RW) should be computed.

dr
A vec specifying on which scales the parameters of a Drift (DR) should be computed.

ad
A vec of the Allan variance.

scales
A vec of the scales.

Value

A list with the estimated parameters.