Summary Allan Variance

Displays the summary table of the output of the `avar()` function

# S3 method for avar
summary(object, ...)

Arguments

object
A avar object.

...
Additional arguments affecting the summary produced. A table that contains:

"Time": The averaging time at each level.

"AVar": The estimated Allan variance.

"ADev": The estimated Allan deviation.

"Lower CI": The lower bound of the confidence interval for the Allan deviation (ADev).

"Upper CI": The upper bound of the confidence interval for the Allan deviation (ADev).

Examples

# Set seed for reproducibility
set.seed(999)

# Simulate a Gaussian white noise
Xt = rnorm(10000)

# Compute Allan
out = avar(Xt)

# Summary
summary(out)

#>       Time         AVar        ADev     Lower CI   Upper CI
#>  [1,]    2 4.952007e-01 0.703705007  0.689629499 0.71778051
#>  [2,]    4 2.502655e-01 0.500265470  0.486112995 0.51441794
#>  [3,]    8 1.271406e-01 0.356567871  0.342299448 0.37083629
#>  [4,]   16 6.336803e-02 0.251730079  0.237478669 0.26598149
#>  [5,]   32 2.946504e-02 0.171653845  0.157899513 0.18540818
#>  [6,]   64 1.434979e-02 0.119790597  0.106194260 0.13338693
#>  [7,]  128 6.268502e-03 0.079173874  0.066424194 0.09192356
#>  [8,]  256 3.270200e-03 0.057185663  0.044077148 0.07029418
#>  [9,]  512 1.645913e-03 0.040569859  0.027241824 0.05389789
#> [10,] 1024 7.058425e-04 0.026567697  0.013877234 0.03925816
#> [11,] 2048 4.983802e-04 0.022324430  0.006302229 0.03834663
#> [12,] 4096 9.023161e-05 0.009499032 -0.001690222 0.02068829
#> attr(,"class")
#> [1] "summary.avar" "matrix"